requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
boto3>=1.34.72,<2.0.0
mysql-connector-python==8.2.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
rank-bm25==0.2.2
pandas==2.1.4
langchain==0.1.0
langchain-aws==0.1.0
arize-phoenix>=4.0.0
openinference-instrumentation-bedrock>=0.1.0
openinference-instrumentation-langchain>=0.1.0
opentelemetry-exporter-otlp>=1.20.0